HZN-1307: Refactor the Camel Sink impl. into different modules, allowing the client (producer) and server (consumer) to be loaded separately.
