--- core/ipc/sink/camel-impl/src/main/resources/META-INF/opennms/applicationContext-ipc-sink-server-camel.xml
@@ -1,53 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<beans xmlns="http://www.springframework.org/schema/beans"
-  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xmlns:tx="http://www.springframework.org/schema/tx"
-  xmlns:context="http://www.springframework.org/schema/context"
-  xmlns:util="http://www.springframework.org/schema/util"
-  xmlns:onmsgi="http://xmlns.opennms.org/xsd/spring/onms-osgi"
-  xsi:schemaLocation="
-  http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans-4.2.xsd
-  http://www.springframework.org/schema/context http://www.springframework.org/schema/context/spring-context-4.2.xsd
-  http://www.springframework.org/schema/tx http://www.springframework.org/schema/tx/spring-tx-4.2.xsd
-  http://www.springframework.org/schema/util http://www.springframework.org/schema/util/spring-util-4.2.xsd
-  http://xmlns.opennms.org/xsd/spring/onms-osgi http://xmlns.opennms.org/xsd/spring/onms-osgi.xsd
-  http://camel.apache.org/schema/spring http://camel.apache.org/schema/spring/camel-spring-2.19.1.xsd
-">
-
-  <context:annotation-config />
-
-  <bean id="identity" class="org.springframework.beans.factory.config.MethodInvokingFactoryBean">
-    <property name="targetObject" ref="distPollerDao"/>
-    <property name="targetMethod" value="whoami"/>
-  </bean>
-
-  <bean id="location" class="org.springframework.beans.factory.config.MethodInvokingFactoryBean">
-    <property name="targetObject" ref="identity"/>
-    <property name="targetMethod" value="getLocation"/>
-  </bean>
-
-  <bean id="camelMessageConsumerManager" class="org.opennms.core.ipc.sink.camel.CamelMessageConsumerManager" >
-    <constructor-arg ref="sinkServer"/>
-  </bean>
-
-  <onmsgi:service ref="camelMessageConsumerManager" interface="org.opennms.core.ipc.sink.api.MessageConsumerManager" />
-
-  <bean id="camelLocalMessageDispatcherFactory" class="org.opennms.core.ipc.sink.camel.CamelLocalMessageDispatcherFactory" />
-
-  <onmsgi:service ref="camelLocalMessageDispatcherFactory" interface="org.opennms.core.ipc.sink.api.MessageDispatcherFactory" />
-
-  <!-- Reduces the graceful shutdown time from 300 to 15 seconds. -->
-  <bean id="shutdownStrategy" class="org.apache.camel.impl.DefaultShutdownStrategy">
-    <property name="timeout" value="15"/>
-  </bean>
-
-  <camelContext id="sinkServer" xmlns="http://camel.apache.org/schema/spring">
-    <!-- Exchanges contain credentials i.e. SNMP community details, so avoid logging them -->
-    <properties>
-      <property key="CamelLogDebugBodyMaxChars" value="{{sys:org.opennms.ipc.camel.body.debug:-1}}"/>
-    </properties>
-
-    <!-- The routes are managed by the CamelMessageConsumerManager -->
-  </camelContext>
-
-</beans>
