+++ core/ipc/sink/camel/client/src/main/resources/OSGI-INF/blueprint/blueprint-ipc-client.xml
@@ -1,0 +1,48 @@
+<blueprint xmlns="http://www.osgi.org/xmlns/blueprint/v1.0.0"
+    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+    xmlns:cm="http://aries.apache.org/blueprint/xmlns/blueprint-cm/v1.3.0"
+    xmlns:ext="http://aries.apache.org/blueprint/xmlns/blueprint-ext/v1.5.0"
+    xsi:schemaLocation="
+		http://www.osgi.org/xmlns/blueprint/v1.0.0 
+		https://osgi.org/xmlns/blueprint/v1.0.0/blueprint.xsd
+		http://aries.apache.org/blueprint/xmlns/blueprint-cm/v1.3.0
+		http://aries.apache.org/schemas/blueprint-cm/blueprint-cm-1.3.0.xsd
+		http://aries.apache.org/blueprint/xmlns/blueprint-ext/v1.5.0
+		http://aries.apache.org/schemas/blueprint-ext/blueprint-ext-1.5.xsd
+		http://camel.apache.org/schema/blueprint
+		http://camel.apache.org/schema/blueprint/camel-blueprint-2.19.1.xsd
+">
+
+    <cm:property-placeholder id="ipcProperties"
+      persistent-id="org.opennms.core.ipc" update-strategy="reload">
+      <cm:default-properties>
+        <cm:property name="body.debug" value="-1" />
+      </cm:default-properties>
+    </cm:property-placeholder>
+
+    <!-- Provided by minion-core -->
+    <reference id="minionIdentity" interface="org.opennms.minion.core.api.MinionIdentity" />
+
+    <!-- Provided by minion-core -->
+    <reference id="queuingservice" interface="org.apache.camel.Component" filter="(alias=opennms.broker)" />
+
+    <bean id="camelRemoteMessageDispatcherFactory" class="org.opennms.core.ipc.sink.camel.client.CamelRemoteMessageDispatcherFactory"
+      init-method="registerJmxReporter" destroy-method="unregisterJmxReporter" />
+    <service ref="camelRemoteMessageDispatcherFactory" interface="org.opennms.core.ipc.sink.api.MessageDispatcherFactory"/>
+
+    <camelContext id="sinkClient" xmlns="http://camel.apache.org/schema/blueprint">
+      <!-- Exchanges contain credentials i.e. SNMP community details, so avoid logging them -->
+      <properties>
+        <property key="CamelLogDebugBodyMaxChars" value="{{body.debug}}"/>
+      </properties>
+
+      <route id="sinkClient">
+        <from uri="direct:sendMessage" />
+        <setExchangePattern pattern="InOnly"/>
+        <recipientList>
+          <simple>queuingservice:${header.JmsQueueName}</simple>
+        </recipientList>
+      </route>
+    </camelContext>
+
+</blueprint>
