+++ core/ipc/sink/camel/itests/src/test/java/org/opennms/core/ipc/sink/camel/itests/SinkBlueprintMessageFailureIT.java
@@ -1,0 +1,147 @@
+/*******************************************************************************
+ * This file is part of OpenNMS(R).
+ *
+ * Copyright (C) 2016-2018 The OpenNMS Group, Inc.
+ * OpenNMS(R) is Copyright (C) 1999-2018 The OpenNMS Group, Inc.
+ *
+ * OpenNMS(R) is a registered trademark of The OpenNMS Group, Inc.
+ *
+ * OpenNMS(R) is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published
+ * by the Free Software Foundation, either version 3 of the License,
+ * or (at your option) any later version.
+ *
+ * OpenNMS(R) is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with OpenNMS(R).  If not, see:
+ *      http://www.gnu.org/licenses/
+ *
+ * For more information contact:
+ *     OpenNMS(R) Licensing <license@opennms.org>
+ *     http://www.opennms.org/
+ *     http://www.opennms.com/
+ *******************************************************************************/
+
+package org.opennms.core.ipc.sink.camel.itests;
+
+import java.util.Dictionary;
+import java.util.Map;
+import java.util.Properties;
+import java.util.concurrent.CountDownLatch;
+
+import org.apache.camel.Component;
+import org.apache.camel.util.KeyValueHolder;
+import org.junit.ClassRule;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.opennms.core.ipc.sink.api.MessageConsumer;
+import org.opennms.core.ipc.sink.api.MessageConsumerManager;
+import org.opennms.core.ipc.sink.api.MessageDispatcherFactory;
+import org.opennms.core.ipc.sink.api.SinkModule;
+import org.opennms.core.ipc.sink.api.SyncDispatcher;
+import org.opennms.core.ipc.sink.camel.itests.heartbeat.Heartbeat;
+import org.opennms.core.ipc.sink.camel.itests.heartbeat.HeartbeatModule;
+import org.opennms.core.test.Level;
+import org.opennms.core.test.MockLogAppender;
+import org.opennms.core.test.OpenNMSJUnit4ClassRunner;
+import org.opennms.core.test.activemq.ActiveMQBroker;
+import org.opennms.core.test.camel.CamelBlueprintTest;
+import org.opennms.core.utils.SystemInfoUtils;
+import org.opennms.minion.core.api.MinionIdentity;
+import org.opennms.test.JUnitConfigurationEnvironment;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.test.context.ContextConfiguration;
+
+@RunWith(OpenNMSJUnit4ClassRunner.class)
+@ContextConfiguration(locations={
+        "classpath:/META-INF/opennms/applicationContext-soa.xml",
+        "classpath:/META-INF/opennms/applicationContext-mockDao.xml",
+        "classpath:/META-INF/opennms/applicationContext-proxy-snmp.xml",
+        "classpath:/META-INF/opennms/applicationContext-queuingservice-mq-vm.xml",
+        "classpath:/META-INF/opennms/applicationContext-ipc-sink-camel-server.xml",
+        "classpath:/META-INF/opennms/applicationContext-ipc-sink-camel-client.xml"
+})
+@JUnitConfigurationEnvironment
+public class SinkBlueprintMessageFailureIT extends CamelBlueprintTest {
+
+    private static final String REMOTE_LOCATION_NAME = "remote";
+
+    @ClassRule
+    public static ActiveMQBroker broker = new ActiveMQBroker();
+
+    @Autowired
+    @Qualifier("queuingservice")
+    private Component queuingservice;
+
+    @Autowired
+    private MessageConsumerManager consumerManager;
+    
+    @SuppressWarnings( "rawtypes" )
+    @Override
+    protected void addServicesOnStartup(Map<String, KeyValueHolder<Object, Dictionary>> services) {
+        services.put(MinionIdentity.class.getName(),
+                new KeyValueHolder<Object, Dictionary>(new MinionIdentity() {
+                    @Override
+                    public String getId() {
+                        return "0";
+                    }
+                    @Override
+                    public String getLocation() {
+                        return REMOTE_LOCATION_NAME;
+                    }
+                }, new Properties()));
+
+        Properties props = new Properties();
+        props.setProperty("alias", "opennms.broker");
+        services.put(Component.class.getName(), new KeyValueHolder<Object, Dictionary>(queuingservice, props));
+    }
+
+    @Override
+    protected String getBlueprintDescriptor() {
+        return "classpath:/OSGI-INF/blueprint/blueprint-ipc-client.xml";
+    }
+
+    @Test(timeout=60000)
+    public void doesntLogMessageBody() throws Exception {
+        HeartbeatModule module = new HeartbeatModule();
+
+        final CountDownLatch consumed = new CountDownLatch(1);
+        // Create a consumer that just throws exceptions during the exchange
+        MessageConsumer<Heartbeat, Heartbeat> consumer = new MessageConsumer<Heartbeat, Heartbeat>() {
+            @Override
+            public SinkModule<Heartbeat, Heartbeat> getModule() {
+                return module;
+            }
+
+            @Override
+            public void handleMessage(Heartbeat heartbeat) {
+                consumed.countDown();
+                throw new IllegalStateException();
+            }
+        };
+        consumerManager.registerConsumer(consumer);
+
+        // Fetch the remote dispatcher from the blueprint context
+        MessageDispatcherFactory remoteMessageDispatcherFactory = context.getRegistry().lookupByNameAndType("camelRemoteMessageDispatcherFactory", MessageDispatcherFactory.class);
+        SyncDispatcher<Heartbeat> dispatcher = remoteMessageDispatcherFactory.createSyncDispatcher(HeartbeatModule.INSTANCE);
+
+        dispatcher.send(new Heartbeat());
+        consumed.await();
+
+        // Sleep slightly longer to allow the body to be logged on the sink consumer listener thread
+        Thread.sleep(2000);
+
+        // Verify that the exchange error was logged
+        MockLogAppender.assertLogMatched(Level.ERROR, "Message History");
+        MockLogAppender.assertLogMatched(Level.ERROR, "queuingservice://" + SystemInfoUtils.DEFAULT_INSTANCE_ID + ".Sink.Heartbeat");
+        // Verify that the message body was suppressed
+        MockLogAppender.assertNoLogMatched(Level.ERROR, "<heartbeat/>");
+
+        consumerManager.unregisterConsumer(consumer);
+    }
+}
