+++ core/ipc/sink/camel/pom.xml
@@ -1,22 +1,19 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
   <parent>
-    <groupId>org.opennms.core</groupId>
-    <artifactId>org.opennms.core.ipc</artifactId>
+    <groupId>org.opennms.core.ipc</groupId>
+    <artifactId>org.opennms.core.ipc.sink</artifactId>
     <version>22.0.0-SNAPSHOT</version>
   </parent>
   <modelVersion>4.0.0</modelVersion>
-  <groupId>org.opennms.core.ipc</groupId>
-  <artifactId>org.opennms.core.ipc.sink</artifactId>
+  <groupId>org.opennms.core.ipc.sink</groupId>
+  <artifactId>org.opennms.core.ipc.sink.camel</artifactId>
   <packaging>pom</packaging>
-  <name>OpenNMS :: Core :: IPC :: Sink</name>
+  <name>OpenNMS :: Core :: IPC :: Sink :: Camel</name>
   <modules>
-    <module>api</module>
     <module>common</module>
-    <module>camel</module>
-    <module>kafka-impl</module>
-    <module>aws-sqs-impl</module>
-    <module>mock-impl</module>
-    <module>xml</module>
+    <module>client</module>
+    <module>server</module>
+    <module>itests</module>
   </modules>
 </project>
