+++ core/ipc/sink/camel/server/src/main/resources/META-INF/opennms/component-dao.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<beans xmlns="http://www.springframework.org/schema/beans" 
+       xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+       xsi:schemaLocation="
+       http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd
+       ">
+
+    <!-- Conditionally load the Camel Sink -->
+    <bean class="org.opennms.core.ipc.sink.camel.server.ConditionalCamelSinkContext"/>
+
+</beans>
