+++ core/ipc/sink/camel/server/src/main/resources/OSGI-INF/blueprint/blueprint-ipc-server.xml
@@ -1,0 +1,37 @@
+<blueprint xmlns="http://www.osgi.org/xmlns/blueprint/v1.0.0"
+    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+    xmlns:cm="http://aries.apache.org/blueprint/xmlns/blueprint-cm/v1.3.0"
+    xsi:schemaLocation="
+		http://www.osgi.org/xmlns/blueprint/v1.0.0 
+		https://osgi.org/xmlns/blueprint/v1.0.0/blueprint.xsd
+		http://aries.apache.org/blueprint/xmlns/blueprint-cm/v1.3.0
+		http://aries.apache.org/schemas/blueprint-cm/blueprint-cm-1.3.0.xsd
+		http://camel.apache.org/schema/blueprint
+		http://camel.apache.org/schema/blueprint/camel-blueprint-2.19.1.xsd
+">
+
+    <cm:property-placeholder id="ipcProperties"
+      persistent-id="org.opennms.core.ipc" update-strategy="reload">
+      <cm:default-properties>
+        <cm:property name="body.debug" value="-1" />
+      </cm:default-properties>
+    </cm:property-placeholder>
+
+    <reference id="queuingservice" interface="org.apache.camel.Component" filter="(alias=opennms.broker)" />
+
+    <bean id="camelMessageConsumerManager" class="org.opennms.core.ipc.sink.camel.server.CamelMessageConsumerManager" >
+        <argument ref="sinkServer"/>
+    </bean>
+
+    <service ref="camelMessageConsumerManager" interface="org.opennms.core.ipc.sink.api.MessageConsumerManager" />
+
+    <camelContext id="sinkServer" xmlns="http://camel.apache.org/schema/blueprint">
+        <!-- Exchanges contain credentials i.e. SNMP community details, so avoid logging them -->
+        <properties>
+            <property key="CamelLogDebugBodyMaxChars" value="{{body.debug}}"/>
+        </properties>
+
+        <!-- The routes are managed by the CamelMessageConsumerManager -->
+    </camelContext>
+
+</blueprint>
